zaakafhandelcomponent-1749 Policies nalopen omtrent toekennen/verdelen van zaken

+++ src/main/java/net/atos/zac/app/zaken/ZakenRESTService.java
@@ -23,6 +23,7 @@ import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -481,48 +482,34 @@ public class ZakenRESTService {
         return policyService.filterAllowedZaaktypen(zaaktypen).stream().map(zaaktypeConverter::convert).toList();
     }
 
-    @PUT
+    @PATCH
     @Path("toekennen")
     public RESTZaak toekennen(final RESTZaakToekennenGegevens toekennenGegevens) {
         final Zaak zaak = zrcClientService.readZaak(toekennenGegevens.zaakUUID);
         assertPolicy(zaak.isOpen() && policyService.readZaakRechten(zaak).getToekennen());
-        if (StringUtils.isNotEmpty(toekennenGegevens.behandelaarGebruikersnaam)) {
-            // Toekennen of overdragen
-            final User user = identityService.readUser(toekennenGegevens.behandelaarGebruikersnaam);
-            zrcClientService.updateRol(zaak, bepaalRolMedewerker(user, zaak), toekennenGegevens.reden);
-        } else {
-            // Vrijgeven
-            zrcClientService.deleteRol(zaak, BetrokkeneType.MEDEWERKER, toekennenGegevens.reden);
+
+        final RolMedewerker rolBehandelaar = zgwApiService.findBehandelaarForZaak(zaak);
+        String behandelaar = rolBehandelaar != null ? rolBehandelaar.getBetrokkeneIdentificatie().getIdentificatie() : null;
+
+        final RolOrganisatorischeEenheid groep = zgwApiService.findGroepForZaak(zaak);
+
+        if (!Objects.equals(behandelaar, toekennenGegevens.behandelaarGebruikersnaam)) {
+            if (StringUtils.isNotEmpty(toekennenGegevens.behandelaarGebruikersnaam)) {
+                // Toekennen of overdragen
+                final User user = identityService.readUser(toekennenGegevens.behandelaarGebruikersnaam);
+                zrcClientService.updateRol(zaak, bepaalRolMedewerker(user, zaak), toekennenGegevens.reden);
+            } else {
+                // Vrijgeven
+                zrcClientService.deleteRol(zaak, BetrokkeneType.MEDEWERKER, toekennenGegevens.reden);
+            }
         }
 
-        if (StringUtils.isNotEmpty(toekennenGegevens.groepId)) {
-            groepToekennen(zaak, toekennenGegevens);
+        if (!Objects.equals(groep.getBetrokkeneIdentificatie().getIdentificatie(), toekennenGegevens.groepId)) {
+            final Group group = identityService.readGroup(toekennenGegevens.groepId);
+            zrcClientService.updateRol(zaak, bepaalRolGroep(group, zaak), toekennenGegevens.reden);
         }
+
         return zaakConverter.convert(zaak);
-    }
-
-    @PUT
-    @Path("toekennen/groep")
-    public RESTZaak groepToekennen(final RESTZaakToekennenGegevens toekennenGegevens) {
-        final Zaak zaak = zrcClientService.readZaak(toekennenGegevens.zaakUUID);
-        groepToekennen(zaak, toekennenGegevens);
-        return zaakConverter.convert(zaak);
-    }
-
-    private void groepToekennen(final Zaak zaak, final RESTZaakToekennenGegevens toekennenGegevens) {
-        assertPolicy(zaak.isOpen() && policyService.readZaakRechten(zaak).getToekennen());
-        final Group group = identityService.readGroup(toekennenGegevens.groepId);
-        zrcClientService.updateRol(zaak, bepaalRolGroep(group, zaak), toekennenGegevens.reden);
-    }
-
-    @PUT
-    @Path("vrijgeven")
-    public void vrijgeven(final RESTZaakToekennenGegevens toekennenGegevens) {
-        final Zaak zaak = zrcClientService.readZaak(toekennenGegevens.zaakUUID);
-        assertPolicy(zaak.isOpen() && policyService.readZaakRechten(zaak).getToekennen());
-
-        zrcClientService.deleteRol(zaak, BetrokkeneType.MEDEWERKER, toekennenGegevens.reden);
-        indexeerService.indexeerDirect(toekennenGegevens.zaakUUID.toString(), ZoekObjectType.ZAAK);
     }
 
     @PUT
@@ -617,9 +604,7 @@ public class ZakenRESTService {
     @PUT
     @Path("toekennen/mij")
     public RESTZaak toekennenAanIngelogdeMedewerker(final RESTZaakToekennenGegevens toekennenGegevens) {
-        final Zaak zaak = zrcClientService.readZaak(toekennenGegevens.zaakUUID);
-        assertPolicy(zaak.isOpen() && policyService.readZaakRechten(zaak).getToekennen());
-        ingelogdeMedewerkerToekennenAanZaak(zaak, toekennenGegevens);
+        final Zaak zaak = ingelogdeMedewerkerToekennenAanZaak(toekennenGegevens);
         return zaakConverter.convert(zaak);
     }
 
@@ -627,8 +612,7 @@ public class ZakenRESTService {
     @Path("toekennen/mij/lijst")
     public RESTZaakOverzicht toekennenAanIngelogdeMedewerkerVanuitLijst(final RESTZaakToekennenGegevens toekennenGegevens) {
         assertPolicy(policyService.readWerklijstRechten().getZakenTaken());
-        final Zaak zaak = zrcClientService.readZaak(toekennenGegevens.zaakUUID);
-        ingelogdeMedewerkerToekennenAanZaak(zaak, toekennenGegevens);
+        final Zaak zaak = ingelogdeMedewerkerToekennenAanZaak(toekennenGegevens);
         indexeerService.indexeerDirect(zaak.getUuid().toString(), ZoekObjectType.ZAAK);
         return zaakOverzichtConverter.convert(zaak);
     }
@@ -746,9 +730,13 @@ public class ZakenRESTService {
         return resultaattypeConverter.convertResultaattypes(ztcClientService.readResultaattypen(ztcClientService.readZaaktype(zaaktypeUUID).getUrl()));
     }
 
-    private void ingelogdeMedewerkerToekennenAanZaak(final Zaak zaak, final RESTZaakToekennenGegevens toekennenGegevens) {
+    private Zaak ingelogdeMedewerkerToekennenAanZaak(final RESTZaakToekennenGegevens toekennenGegevens) {
+        final Zaak zaak = zrcClientService.readZaak(toekennenGegevens.zaakUUID);
+        assertPolicy(zaak.isOpen() && policyService.readZaakRechten(zaak).getToekennen());
+
         final User user = identityService.readUser(loggedInUserInstance.get().getId());
         zrcClientService.updateRol(zaak, bepaalRolMedewerker(user, zaak), toekennenGegevens.reden);
+        return zaak;
     }
 
     private RolOrganisatorischeEenheid bepaalRolGroep(final Group group, final Zaak zaak) {
